mediaplayer test component and application allows instantiation
