+++ core/src/main/java/com/guardian/core/dagger/CoreComponent.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import com.guardian.core.feed.FeedRepository
 import com.guardian.core.feeditem.FeedItemRepository
 import com.guardian.core.mediametadata.MediaMetadataRepository
+import com.guardian.core.mediaplayer.PackageValidator
 import com.guardian.core.mediaplayer.common.MediaSessionConnection
 import com.guardian.core.mediaplayer.podx.PodXEventEmitter
 import com.guardian.core.podxevent.PodXEventRepository
@@ -20,6 +21,7 @@ import javax.inject.Singleton
     SearchDataModule::class,
     RepositoryModule::class,
     MediaSessionConnectionModule::class,
+    MediaServiceModule::class,
     RoomModule::class])
 interface CoreComponent {
 
@@ -37,4 +39,5 @@ interface CoreComponent {
     fun provideMediaMetadataRepository(): MediaMetadataRepository
     fun provideMediaSessionConnection(): MediaSessionConnection
     fun providePodXEventEmitter(): PodXEventEmitter
+    fun providePackageValidator(): PackageValidator
 }
