--- core/src/test/java/com/guardian/core/TestCoreComponent.kt
@@ -1,19 +1,0 @@
-package com.guardian.core
-
-import com.guardian.core.dagger.CoreComponent
-import dagger.Component
-import dagger.android.AndroidInjectionModule
-
-@TestScope
-@Component(dependencies = [CoreComponent::class],
-    modules = [ServicesForTestingModule::class,
-        AndroidInjectionModule::class])
-interface TestCoreComponent {
-    @Component.Builder
-    interface Builder {
-        fun coreComponent(coreComponent: CoreComponent): Builder
-        fun build(): TestCoreComponent
-    }
-
-    fun inject(coreComponentTestApplication: CoreComponentTestApplication)
-}
