+++ core/src/test/java/com/guardian/core/mediaplayer/daggermocks/TestPackageValidator.kt
@@ -1,0 +1,9 @@
+package com.guardian.core.mediaplayer.daggermocks
+
+import com.guardian.core.mediaplayer.PackageValidator
+
+class TestPackageValidator : PackageValidator {
+    override fun isKnownCaller(callingPackage: String, callingUid: Int): Boolean {
+        return true
+    }
+}
